Fixed problem with opening save archive
For some (still unknown) reason, the program needs another fs:USER
handle and needs to override the normal/default one with it by using
fsUseSession(). Once this is done, the program behaves as expected.

Took the opportunity to use some of the new result.h macros to check
result values, and also did some cleaning.